Add files via upload
At the company I work for, we need a report with several XML files from the client, invoices that he needed to do his accounting. So I tried to make a query through his database, but it was unfeasible and very poorly optimized because it was Firebird.

So I made a small script that runs in the console, even in Python, and that's it.